requirements.txt
# Core dependencies
python-dotenv==1.0.1
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# RAG and Embeddings
faiss-cpu==1.7.4
sentence-transformers==2.5.1
rank-bm25==0.2.2
numpy==1.24.3
scikit-learn==1.3.2

# LLM and Training
torch==2.2.1
transformers==4.37.2
datasets==2.17.1
peft==0.9.0
accelerate==0.27.2
bitsandbytes==0.42.0
langchain==0.1.11
langchain-community==0.0.27

# Utilities
tiktoken==0.6.0
tqdm==4.66.2

# Database and Caching
redis==5.0.2

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
httpx==0.27.0 